Move specialitiesMap, servicesList and servicesRegExp into a helper file
