+++ src/components/Header/TopNav/ServicesSpecialitiesMap.js
@@ -1,0 +1,14 @@
+const specialitiesMap = {
+  engineering: ['node-js', 'graphql', 'vue-js', 'react-js', 'kubernetes'],
+  design: [],
+  training: [],
+  delivery: [],
+  dedicatedTeams: [],
+  openSource: []
+}
+
+const servicesList = Object.keys(specialitiesMap)
+
+const servicesRegExp = new RegExp(servicesList.join('|'))
+
+export { specialitiesMap, servicesList, servicesRegExp }
